Update main.py
Hey Isaac. This is what we got done during the rest of the meeting. So it greets the user and we already added a button that takes you to the next page where it has two images. You can possibly work on selecting one of the two images and making a score counter.